add pipelines
