+++ .gitignore
@@ -2,3 +2,4 @@
 flamegraph.svg
 *.mp4
 *.mp3
+.env
