+++ astro/src/quadtree.rs
@@ -146,7 +146,7 @@ where
             let r = (location[0] - self.centre[0]).powi(2)
                 + (location[1] - self.centre[1]).powi(2)
                 + (location[2] - self.centre[2]).powi(2).sqrt();
-            if r / self.extent < bh_factor {
+            if self.extent / r < bh_factor {
                 return vec![e];
             }
         }
