+++ readme.md
@@ -1,34 +1,72 @@
 # Physim
-A program for performing many-body simulations.
+An extensible framework for performing N-body simulations.
 
-## Design goals:
-The aim is to make a program the functions similar to gstreamer.
-- CLI interface. The interface should work like 
-  - `physim plummer n=1000 ! barneshut theta=10.0 ! video`
-  - `physim cube n=1000 seed=2 ! exact ! video`
-  - `physim spiral ! barneshut ! stdout`
-  - `physim file location="stars.csv" ! fileparse units="units.txt" ! barneshut ! stdout`  
-  - It should have autocomple and help for each element. Each element should have parameters which can be passed to it.
-- Plugin style architecture. This will be acheived with dynamic library loading. Plugins will provide elements that can be combined to produce different simulations. These plugins should be writable in any language, but specically in Rust, C, and Go.
-- A renderer which supports camera movements.
-- Can output to ffmpeg for creation of nice videos.
-- Events which can affect the simulation e.g. create new entities on click, terminate the simulation gracefully.
-- Be highly performant.
+## Overview
+Physim provides a framework for users to run N-body simulations. Users can build pipelines that define the properties of a simulation and have access to a variety of useful elements such as OpenGL renderers and gravity calculations.
 
-## Creating videos with the stdout element
-Note, CLI program a WIP.
+The functionality of physim can be expanded with plugins to add functionality. Developers can make these plugins with `Rust` and there is support for some elements to be written in languages with a `C` ABI.
+
+### Roadmap
+There are many features that are under development or planned. The library should be considered unstable and backward compatibility is not guaranteed.
+
+The current state of the framework is `pre-0.1.0`.
+
+#### 0.1.0
+This is the proof-of-concept release. It'll demonstrate core concepts. There will be no patch versions and bugs discovered in this release will be fixed in `0.2.0`.
+- `physim`: CLI program for running simulations.
+- `physcan`: CLI program for getting documentation about elements
+- Plugin loading via dynamic library loading.
+- Transform API with a C ABI.
+- Render API.
+- Initialiser API.
+- `glrender` plugin. Provides the `glrender` and `stdout` elements.
+- `astro` plugin. Provides assorted elements for galaxy simulations.
+- Rust macros for creating plugins.
+#### 0.2.0
+This release will have more advanced elements as well as extend the functionality of the current elements. There will be no patch versions and bugs discovered in this release will be fixed in `0.3.0`.
+- Inter-element communication bus. Include events such as "end-of-simulation" to enable clean finishes and dynamic updating of the properties of elements.
+- Global simulation parameters like the time step, number of steps and "universe bounds" will be configurable.
+- Transmute API. This will support events like entity mergers.
+- Synthesis API. This will support events such as entity creation from external inputs.
+- `glrender` elements will have more properties exposed to the user.
+- Support multiple transform elements in one pipeline.
+- Add more elements to the `astro` plugin. The `plummer` element will be added in this release.
+#### 0.3.0
+This release will focus on performance, testability, bug fixing and stability along with expanding the functionality provided by the `astro` plugin. Things included in this release will be.
+- demonstration plugins written in `C` and `Go`.
+- The physical unit system will be properly established in the `astro` element.
+- `glrender` with more advanced user inputs.
+
+#### Beyond.
+No further features are planned yet. A gas simulation plugin would be interesting as would some electromagnetism simulations. Adding a simulation with relativistic effects might be possible with this framework. These features could result in `Entity` changing, so it is worth reiterating that this library is not in a stable form and backwards compatibility is not guarenteed.
+
+
+## Running a simulation
+
+`physim` simulations can be configured directly in the CLI. Each element is delimited by `!`, and the properties of the element can be configured as shown in the following example:
+
 ```
-cargo run -r --bin physim | ffmpeg -y -f rawvideo -pixel_format bgra -video_size 1920x1080 -framerate 60 -i pipe:0  -c:v libx265 -preset fast -crf 23 -pix_fmt yuv420p output.mp4
+cargo run -r --bin physim cube n=50000 seed=1 ! star mass=10000.0 x=0.5 y=0.5 radius=0.1 ! star mass=10000.0 x=-0.5 y=-0.5 radius=0.1 ! astro theta=1.5 ! glrender
 ```
-## State management
-- elements should  have properties.
-- when you want to use an element, you should need to lock it and be safe to read its state.
-- setting state should be somehow generic.
-- stretch goal: add macro to derive documentation.
+
+### Creating videos with the stdout element
+The following example shows how to use the `stdout` element from the `glrender` plugin.
 ```
-let e = element.lock()
-e.set_props(hashmap)
+cargo run -r --bin physim cube n=50000 seed=1 ! star mass=10000.0 x=0.5 y=0.5 radius=0.1 ! star mass=10000.0 x=-0.5 y=-0.5 radius=0.1 ! astro theta=1.5 ! stdout | ffmpeg -y -f rawvideo -pixel_format bgra -video_size 1920x1080 -framerate 60 -i pipe:0  -c:v libx265 -preset fast -crf 23 -pix_fmt yuv420p output.mp4
 ```
+### Element documentation
+Specify the location of your plugin directory with the `PHYSIM_PLUGIN_DIR` environment variable. To determine what elements you have access to, run
+```
+cargo run -r --bin physcan
+```
+and to see more details about an element:
+```
+cargo run -r --bin physcan cube
+```
+
+## Development
+
+TODO: create a plugin tutorial.
 
 ## Licence
 MIT.
